# PointValueStoreClient/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PointValueStoreConnector",
    "version": "1.0"
  },
  "paths": {
    "/points/{pointId}/values": {
      "get": {
        "tags": [
          "Points"
        ],
        "parameters": [
          {
            "name": "pointId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "upTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseValueContainerDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseValueContainerDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseValueContainerDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Points"
        ],
        "parameters": [
          {
            "name": "pointId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValueContainerDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ValueContainerDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ValueContainerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "NumericValueDto": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "value": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ResponseValueContainerDto": {
        "type": "object",
        "properties": {
          "values": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NumericValueDto"
            },
            "nullable": true
          },
          "unit": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ValueContainerDto": {
        "type": "object",
        "properties": {
          "values": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NumericValueDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}